FindDirection: Adds special case handling for identical elements
If all the amplitudes are identical, there is no logical place for the DF to
point. We may already be at the local max of light intensity and all elements
are equidistant from the light source, or perhaps we are so far away that all
elements are dark. Either way, we don't have enough info to make a movement
decision so stay here.